refactor: improve types
