+++ packages/swapper/src/index.ts
@@ -4,3 +4,4 @@ export * from './constants'
 export * from './swapper'
 export * from './thorchain-utils'
 export * from './safe-utils'
+export * from './swappers/CowSwapper'
