+++ packages/swapper/src/swappers/CowSwapper/utils/constants.ts
@@ -9,29 +9,6 @@ export const DEFAULT_ADDRESS = zeroAddress
 export const COW_SWAP_VAULT_RELAYER_ADDRESS = '0xc92e8bdf79f0507f65a392b0ab4667716bfe0110'
 export const COW_SWAP_SETTLEMENT_ADDRESS = '0x9008D19f58AAbD9eD0D60971565AA8510560ab41'
 
-export enum CoWSwapOrderKind {
-  Buy = 'buy',
-  Sell = 'sell',
-}
-
-export enum CoWSwapSigningScheme {
-  EIP712 = 'eip712',
-  EIP1271 = 'eip1271',
-  EthSign = 'ethsign',
-  PreSign = 'presign',
-}
-
-export enum CoWSwapSellTokenSource {
-  ERC20 = 'erc20',
-  External = 'external',
-  Internal = 'internal',
-}
-
-export enum CoWSwapBuyTokenDestination {
-  ERC20 = 'erc20',
-  Internal = 'internal',
-}
-
 // Address used by CowSwap to buy ETH
 // See https://github.com/gnosis/gp-v2-contracts/commit/821b5a8da213297b0f7f1d8b17c893c5627020af#diff-12bbbe13cd5cf42d639e34a39d8795021ba40d3ee1e1a8282df652eb161a11d6R13
 export const COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE'
